- Added Redux thunk for household (a copy of account).
