+++ screens/HomeScreen.tsx
@@ -2,24 +2,21 @@ import * as React from 'react';
 import { View, Text, Button, Alert } from 'react-native';
 import { styles } from '../styles/Styles';
 import { useAppSelector, useAppDispatch } from '../store/store';
-import { selectAllAccounts } from '../store/account/accountSelectors';
-import { selectAllHouseholds } from '../store/household/householdSelector';
-import { AddAccount } from '../store/account/accountActions';
+import { selectAllHouseholds } from '../store/hosehold/householdSelectors';
+import { AddHousehold } from '../store/hosehold/householdActions';
 
 function HomeScreen() {
-  const allAccounts = useAppSelector(selectAllAccounts);
   const allHouseholds = useAppSelector(selectAllHouseholds);
   const dispatch = useAppDispatch();
 
   const handleAdd = () => {
-    dispatch(AddAccount({Id: 1, Email: 'a@a.com', Password: '123'}));
+    dispatch(AddHousehold({Id: 1, Name: 'household 1', Code: '123'}));
     Alert.alert('Added new household');
   }
 
   const handlePrint = () => {
     Alert.alert('Print (see console)');
-    // console.log('allAccountsSelector: ', allAccounts);
-    console.log('allHouseholds: ', allAccounts);
+    console.log('allHouseholds: ', allHouseholds);
   }
   return (
     <View style={styles.container}>
