+++ store/hosehold/householdState.ts
@@ -1,0 +1,9 @@
+import { Households } from '../../entities/Household';
+
+export interface HouseholdState {
+  AllHouseholds: Households[];
+}
+
+export const initialState: HouseholdState = {
+  AllHouseholds: []
+}
