--- store/household/householdActions.ts
@@ -1,8 +1,0 @@
-import { Households } from '../../entities/Household';
-
-export interface AddHouseholdAction {
-  type: 'household/addHousehold';
-  payload: Households;
-}
-
-export type KnownActions = AddHouseholdAction;
